Init Commit; First commit with most of code all done.
First commit with most of what is needed already completed. Along with the basic function of saving the RGB value and the hex value to the ListView, I also added a "random" color button and "delete" button used to delete a color from the list. In the future I intend on adding a function to persistently save the data to the built in android database.